Address PR feedback: Renaming and refactoring
