+++ src/aws_credentials_for_environment.py
